requirements.txt
langchain
sentence_transformers
torch
sentencepiece
accelerate
chromadb
pypdf
tiktoken
streamlit
fastapi
uvicorn
python-multipart
aiofiles